refactor: remove repository key typo fixing
Remove support for handling the `repository` key being incorrectly
spelled in the package data object.

BREAKING CHANGE:

Typos for the `repository` key in package data will no longer be
corrected. For example, if the `repository` key is listed under the
`repo`, or `repositories`, name it will no longer be coerced into
`repository` prior to being processed by this package.

Closes #18